Report change events to work better with Posthog
